Rectification toupies chinoises

+++ exerciceP10/general/constantes.cc
@@ -2,4 +2,4 @@
 
 extern const Vecteur g={0.0,0.0, -9.81};
 extern const double eps(1e-8);
-extern const double N(500.0);
+extern const size_t N(500);
