+++ exerciceP10/general/constantes.h
@@ -3,4 +3,4 @@
 
 extern const Vecteur g;  // Vecteur accélération gravité
 extern const double eps; // Epsilon
-extern const double N;   // Nombre de tranche pour construire notre solide de revolution (précision de la construction)
+extern const size_t N;   // Nombre de tranche pour construire notre solide de revolution (précision de la construction)
